requirements.txt
autogen-agentchat
autogen-ext[azure, openai]
python-dotenv
azure-identity
azure-ai-projects


aiohttp
autogen-ext